requirements.txt
fastapi
uvicorn
opencv-python
openai
python-dotenv
numpy
scikit-learn
requests
soundfile
tqdm
python-multipart